+++ packages/builders/src/index.ts
@@ -8,6 +8,7 @@ export { createNodeModuleErrorPlugin } from './node-module-esbuild-plugin.js';
 export { StandaloneBuilder } from './standalone.js';
 export { createSwcPlugin } from './swc-esbuild-plugin.js';
 export type {
+  AstroConfig,
   BuildTarget,
   NextConfig,
   StandaloneConfig,
@@ -15,8 +16,6 @@ export type {
   VercelBuildOutputConfig,
   WorkflowConfig,
 } from './types.js';
-export {
-  isValidBuildTarget,
-  validBuildTargets,
-} from './types.js';
+export { isValidBuildTarget, validBuildTargets } from './types.js';
 export { VercelBuildOutputAPIBuilder } from './vercel-build-output-api.js';
+export { createBuildQueue } from './build-queue.js';